+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Windows: Rename options parameter from "disableAutoGeolocation" to "enableGeolocation" for consistency with Android and iOS.
 - Remove backward compatibility shim for legacy v2 API.
 - Update Typescript types for v4.
+- Support specification of launch mode for Apple Maps on iOS: URI scheme or MapKit class. 
+Default to URI scheme method in order to support address location types without geocoding.  Resolves #106 and #107. 
 
 ##v3.2.2
 - Rename typing files to use NPM package names instead of Github repo names.